first go at getting this working via kind

+++ pipeline/bootstrap-pipeline.sh
@@ -5,9 +5,11 @@ set +e
 run_tkn=${run_tkn:-0}
 skip_tests=${skip_tests:-false}
 GH_TOKEN=${GH_TOKEN}
+ARTIFACTORY_APIKEY=${ARTIFACTORY_APIKEY}
 git_user=${git_user}
+image_source_repo_username=${image_source_repo_username}
 image_repo="${image_repo:-image-registry.openshift-image-registry.svc:5000}"
-
+dockerhub_hostname="${dockerhub_hostname:-wcp-ibm-streams-docker-local.artifactory.swg-devops.com/pipelines-tutorial}"
 helper_text=""
 realpath() {
     [[ $1 = /* ]] && echo "$1" || echo "$PWD/${1#./}"
@@ -30,13 +32,44 @@ if [[ -z "$TMP" ]]; then
     trap cleanup EXIT
 fi
 
+is_openshift="false"
+is_kubernetes="false"
+client=kubectl
+set +e
+kubectl get ns | grep openshift-apiserver
+rc=$?
+if [[ $rc -eq 0 ]]; then
+    is_openshift=true
+    client=oc
+else
+    is_kubernetes=true
+    client=kubectl
+fi
+if [[ ${is_kubernetes} == "true" ]]; then
+    set -e
+    kubectl apply -f ${repo_root}/pipeline/nfs.yaml
+    helm repo add stable https://charts.helm.sh/stable
+    ip=$(kubectl get svc -n default nfs-service -o jsonpath='{.spec.clusterIP}')
+    helm upgrade --install nfs-provisioner stable/nfs-client-provisioner --values ${repo_root}/pipeline/nfs-values.yaml --set nfs.server=${ip} --namespace nfs-provisioner --create-namespace
+fi
+set +e
 rc=1
 if [[ ! -z $1 ]]; then
-    oc get project $1
-    rc=$?
+    if [[ ${is_openshift} == "true" ]]; then
+        oc get project $1
+        rc=$?
+    else
+        kubectl get ns $1
+        rc=$?
+    fi
 else
-    oc get project m4d-system
-    rc=$?
+    if [[ ${is_openshift} == "true" ]]; then
+        oc get project m4d-system
+        rc=$?
+    else
+        kubectl get ns m4d-system
+        rc=$?
+    fi
 fi
 if [[ ! -z $2 ]]; then
     ssh_key=$2
@@ -45,46 +78,83 @@ else
 fi
 set -e
 if [[ $rc -ne 0 ]]; then
-    oc new-project ${1:-m4d-system}
+    if [[ ${is_openshift} == "true" ]]; then
+        oc new-project ${1:-m4d-system}
+    else
+        kubectl create ns ${1:-m4d-system}
+    fi
 else
-    oc project ${1:-m4d-system} 
+    if [[ ${is_openshift} == "true" ]]; then
+        oc project ${1:-m4d-system}
+    else
+        kubectl config set-context --current --namespace=${1:-m4d-system}
+    fi
 fi
 unique_prefix=$(kubectl config view --minify --output 'jsonpath={..namespace}'; echo)
 
 set +e
 # Be smarter about this - just a quick hack for typical default installs
 oc patch storageclass managed-nfs-storage -p '{"metadata": {"annotations": {"storageclass.kubernetes.io/is-default-class": "true"}}}'
+oc patch storageclass standard -p '{"metadata": {"annotations": {"storageclass.kubernetes.io/is-default-class": "false"}}}'
 set -e
 
-oc apply -f ${repo_root}/pipeline/subscription.yaml
-oc apply -f ${repo_root}/pipeline/serverless-subscription.yaml
+if [[ ${is_openshift} == "true" ]]; then
+    oc apply -f ${repo_root}/pipeline/subscription.yaml
+    oc apply -f ${repo_root}/pipeline/serverless-subscription.yaml
 
-cat > ${TMP}/streams_csv_check_script.sh <<EOH
+    cat > ${TMP}/streams_csv_check_script.sh <<EOH
 #!/bin/bash
 set -x
 oc get -n openshift-pipelines csv | grep redhat-openshift-pipelines-operator 
 oc get -n openshift-pipelines csv | grep redhat-openshift-pipelines-operator | grep Succeeded
 EOH
 chmod u+x ${TMP}/streams_csv_check_script.sh
-try_command "${TMP}/streams_csv_check_script.sh"  40 true 5
+    try_command "${TMP}/streams_csv_check_script.sh"  40 true 5
 
-cat > ${TMP}/streams_csv_check_script.sh <<EOH
+    cat > ${TMP}/streams_csv_check_script.sh <<EOH
 #!/bin/bash
 set -x
 oc get -n openshift-operators csv | grep serverless-operator
 oc get -n openshift-operators csv | grep serverless-operator | grep -e Succeeded -e Replacing
 EOH
 chmod u+x ${TMP}/streams_csv_check_script.sh
-try_command "${TMP}/streams_csv_check_script.sh"  40 false 5
+    try_command "${TMP}/streams_csv_check_script.sh"  40 false 5
+    oc apply -f ${repo_root}/pipeline/knative-eventing.yaml
+else
+    kubectl apply --filename https://storage.googleapis.com/tekton-releases/pipeline/latest/release.yaml
+    kubectl apply -f https://github.com/knative/operator/releases/download/v0.15.4/operator.yaml
+    kubectl apply -f https://github.com/knative/eventing/releases/download/v0.21.0/eventing-crds.yaml
+    kubectl apply -f https://github.com/knative/eventing/releases/download/v0.21.0/eventing-core.yaml
+    kubectl wait pod -n tekton-pipelines --all --for=condition=Ready --timeout=3m
+    set +e
+    kubectl create ns knative-eventing
+    set -e
+    cat > ${TMP}/knative-eventing.yaml <<EOH
+apiVersion: operator.knative.dev/v1alpha1
+kind: KnativeEventing
+metadata:
+  name: knative-eventing
+  namespace: knative-eventing
+EOH
+    ls -alrt ${TMP}/
+    kubectl apply -f ${TMP}/knative-eventing.yaml
+fi
 
-oc adm policy add-cluster-role-to-user cluster-admin system:serviceaccount:${unique_prefix}:pipeline
-oc adm policy add-cluster-role-to-user cluster-admin system:serviceaccount:${unique_prefix}:root-sa
-oc adm policy add-role-to-group system:image-puller system:serviceaccounts:${unique_prefix} --namespace ${unique_prefix}
-oc adm policy add-role-to-group system:image-puller system:serviceaccounts:m4d-blueprints --namespace ${unique_prefix}
-oc adm policy add-role-to-user system:image-puller system:serviceaccount:${unique_prefix}:wkc-connector --namespace ${unique_prefix}
-# Temporary hack pending a better solution
-oc adm policy add-scc-to-user anyuid system:serviceaccount:${unique_prefix}:opa-connector
-oc adm policy add-scc-to-user anyuid system:serviceaccount:${unique_prefix}:manager
+if [[ ${is_openshift} == "true" ]]; then
+    oc adm policy add-cluster-role-to-user cluster-admin system:serviceaccount:${unique_prefix}:pipeline
+    oc adm policy add-cluster-role-to-user cluster-admin system:serviceaccount:${unique_prefix}:root-sa
+    oc adm policy add-role-to-group system:image-puller system:serviceaccounts:${unique_prefix} --namespace ${unique_prefix}
+    oc adm policy add-role-to-group system:image-puller system:serviceaccounts:m4d-blueprints --namespace ${unique_prefix}
+    oc adm policy add-role-to-user system:image-puller system:serviceaccount:${unique_prefix}:wkc-connector --namespace ${unique_prefix}
+    # Temporary hack pending a better solution
+    oc adm policy add-scc-to-user anyuid system:serviceaccount:${unique_prefix}:opa-connector
+    oc adm policy add-scc-to-user anyuid system:serviceaccount:${unique_prefix}:manager
+else
+    set +e
+    kubectl create clusterrolebinding ${unique_prefix}-default-cluster-admin --clusterrole=cluster-admin --serviceaccount=${unique_prefix}:default
+    set -e
+    #exit 1
+fi
 
 set +e
 #resource_version=$(oc get -f ${repo_root}/pipeline/make.yaml -o jsonpath='{.metadata.resourceVersion}')
@@ -98,24 +168,25 @@ set -x
 oc apply -f ${repo_root}/pipeline/make.yaml
 oc apply -f ${repo_root}/pipeline/git-clone.yaml
 oc apply -f ${repo_root}/pipeline/buildah.yaml
-oc apply -f ${repo_root}/pipeline/knative-eventing.yaml
-oc apply -f ${repo_root}/pipeline/helm-upgrade-from-repo.yaml
+oc apply -f ${repo_root}/pipeline/skopeo-copy.yaml
 oc apply -f ${repo_root}/pipeline/openshift-client.yaml
+oc apply -f ${repo_root}/pipeline/helm-upgrade-from-source.yaml 
+oc apply -f ${repo_root}/pipeline/helm-upgrade-from-repo.yaml 
 helper_text=""
 
-oc patch clustertask helm-upgrade-from-repo -p '
-[{
-  "op": "replace",
-  "path": "/spec/steps/0/image",
-  "value": "wcp-ibm-streams-docker-local.artifactory.swg-devops.com/pipelines-tutorial/k8s-helm:latest"
-}]' --type=json
-
-oc patch clustertask helm-upgrade-from-source -p '
-[{
-  "op": "replace",
-  "path": "/spec/steps/0/image",
-  "value": "wcp-ibm-streams-docker-local.artifactory.swg-devops.com/pipelines-tutorial/k8s-helm:latest"
-}]' --type=json
+#oc patch clustertask helm-upgrade-from-repo -p '
+#[{
+#  "op": "replace",
+#  "path": "/spec/steps/0/image",
+#  "value": "wcp-ibm-streams-docker-local.artifactory.swg-devops.com/pipelines-tutorial/k8s-helm:latest"
+#}]' --type=json
+#
+#oc patch clustertask helm-upgrade-from-source -p '
+#[{
+#  "op": "replace",
+#  "path": "/spec/steps/0/image",
+#  "value": "wcp-ibm-streams-docker-local.artifactory.swg-devops.com/pipelines-tutorial/k8s-helm:latest"
+#}]' --type=json
 
 set +e
 oc delete -f ${repo_root}/pipeline/pipeline.yaml
@@ -139,18 +210,65 @@ helper_text="If this step fails:
 5. re-run bootstrap.sh
 "
 set -x
- 
-oc get secret -n openshift-config pull-secret -o yaml > ${TMP}/secret.yaml
-helper_text=""
-cp ${TMP}/secret.yaml ${TMP}/secret.yaml.orig
-sed -i.bak "s|namespace: openshift-config|namespace: ${unique_prefix}|g" ${TMP}/secret.yaml
-sed -i.bak "s|name: pull-secret|name: regcred|g" ${TMP}/secret.yaml
-cat ${TMP}/secret.yaml
-oc apply -f ${TMP}/secret.yaml
 
-oc secrets link pipeline regcred --for=mount
-oc secrets link builder regcred --for=mount
-oc secrets link pipeline regcred --for=pull
+set +e 
+oc get secret -n openshift-config pull-secret -o yaml > ${TMP}/secret.yaml
+rc=$?
+if [[ ${rc} -eq 0 ]]; then
+    helper_text=""
+    oc get secret -n openshift-config pull-secret -o=go-template='{{index .data ".dockerconfigjson"}}' | base64 --decode | grep wcp-ibm-streams-docker-local
+    rc=$?
+    if [[ ${rc} -eq 0 ]]; then
+        set -e
+        cp ${TMP}/secret.yaml ${TMP}/secret.yaml.orig
+        sed -i.bak "s|namespace: openshift-config|namespace: ${unique_prefix}|g" ${TMP}/secret.yaml
+        sed -i.bak "s|name: pull-secret|name: regcred|g" ${TMP}/secret.yaml
+        cat ${TMP}/secret.yaml
+        oc apply -f ${TMP}/secret.yaml
+    else
+        if [[ ! -z ${ARTIFACTORY_APIKEY} ]]; then
+            set -e
+            auth=$(echo -n "${image_source_repo_username:-$git_username}:${ARTIFACTORY_APIKEY}" | base64 -w 0)
+            cat > ${TMP}/secret.yaml <<EOH
+{"auths":{"${image_source_repo}":{"username":"${image_source_repo_username:-$git_username}","password":"${ARTIFACTORY_APIKEY}","auth":"${auth}"}}}
+EOH
+            kubectl create secret -n ${unique_prefix} generic regcred --from-file=.dockerconfigjson=${TMP}/secret.yaml --type=kubernetes.io/dockerconfigjson
+        else
+            helper_text="Run the following commands to set up credentials for artifactory:
+
+            export ARTIFACTORY_APIKEY=xxx
+            export image_source_repo_username=user@email.com
+            "
+            exit 1
+        fi
+    fi
+else
+    helper_text=""
+    if [[ ! -z ${ARTIFACTORY_APIKEY} ]]; then
+        set -e
+        auth=$(echo -n "${image_source_repo_username:-$git_username}:${ARTIFACTORY_APIKEY}" | base64 -w 0)
+        cat > ${TMP}/secret.yaml <<EOH
+{"auths":{"${image_source_repo}":{"username":"${image_source_repo_username:-$git_username}","password":"${ARTIFACTORY_APIKEY}","auth":"${auth}"}}}
+EOH
+        kubectl create secret -n ${unique_prefix} generic regcred --from-file=.dockerconfigjson=${TMP}/secret.yaml --type=kubernetes.io/dockerconfigjson
+    else
+        helper_text="Run the following commands to set up credentials for artifactory:
+    
+        export ARTIFACTORY_APIKEY=xxx
+        export image_source_repo_username=user@email.com
+        "
+        exit 1
+    fi
+fi
+
+if [[ ${is_openshift} == "true" ]]; then
+    oc secrets link pipeline regcred --for=mount
+    oc secrets link builder regcred --for=mount
+    oc secrets link pipeline regcred --for=pull
+else
+    kubectl patch serviceaccount default -p '{"imagePullSecrets": [{"name": "regcred"}]}'
+    kubectl patch serviceaccount default -p '{"secrets": [{"name": "regcred"}]}'
+fi
 
 cluster_scoped="false"
 deploy_vault="false"
@@ -177,7 +295,7 @@ if [[ $rc -ne 0 ]]; then
 fi
 
 set +e
-oc get project m4d-system
+oc get ns m4d-system
 rc=$?
 set -e
 if [[ $rc -ne 0 ]]; then
@@ -189,7 +307,25 @@ fi
 oc apply -f ${repo_root}/pipeline/rootsa.yaml
 oc apply -f ${repo_root}/pipeline/statefulset.yaml
 oc apply -f ${repo_root}/pipeline/pvc.yaml
-oc adm policy add-scc-to-user privileged system:serviceaccount:${unique_prefix}:root-sa
+if [[ ${is_openshift} == "true" ]]; then
+    oc adm policy add-scc-to-user privileged system:serviceaccount:${unique_prefix}:root-sa
+fi
+
+pushd ${TMP}
+wget https://storage.googleapis.com/tekton-releases/triggers/latest/release.yaml
+if [[ ${is_openshift} == "true" ]]; then
+    sed -i.bak 's|namespace: tekton-pipelines|namespace: openshift-pipelines|g' ${TMP}/release.yaml
+fi
+cat ${TMP}/release.yaml
+wget https://storage.googleapis.com/tekton-releases/triggers/latest/interceptors.yaml
+if [[ ${is_openshift} == "true" ]]; then
+    sed -i.bak 's|namespace: tekton-pipelines|namespace: openshift-pipelines|g' ${TMP}/interceptors.yaml
+fi
+cat ${TMP}/interceptors.yaml
+popd
+
+oc apply -f ${TMP}/release.yaml
+oc apply -f ${TMP}/interceptors.yaml
 
 oc apply -f ${repo_root}/pipeline/eventlistener/generic-image-pipeline.yaml
 oc apply -f ${repo_root}/pipeline/eventlistener/generic-triggerbinding.yaml
@@ -198,18 +334,6 @@ oc apply -f ${repo_root}/pipeline/eventlistener/generic-watcher-apiserversource.
 oc apply -f ${repo_root}/pipeline/eventlistener/generic-watcher-role.yaml
 oc apply -f ${repo_root}/pipeline/eventlistener/generic-watcher-serviceaccount.yaml
 oc apply -f ${repo_root}/pipeline/eventlistener/print-generic-task.yaml
-
-pushd ${TMP}
-wget https://storage.googleapis.com/tekton-releases/triggers/latest/release.yaml
-sed -i.bak 's|namespace: tekton-pipelines|namespace: openshift-pipelines|g' ${TMP}/release.yaml
-cat ${TMP}/release.yaml
-wget https://storage.googleapis.com/tekton-releases/triggers/latest/interceptors.yaml
-sed -i.bak 's|namespace: tekton-pipelines|namespace: openshift-pipelines|g' ${TMP}/interceptors.yaml
-cat ${TMP}/interceptors.yaml
-popd
-
-oc apply -f ${TMP}/release.yaml
-oc apply -f ${TMP}/interceptors.yaml
 
 set +x
 helper_text="If this step fails, run again - knative related pods may be restarting and unable to process the webhook
@@ -228,7 +352,7 @@ oc delete secret git-token
 set -e
 
 if [[ -z ${GH_TOKEN} ]]; then
-    cat ~/.ssh/known_hosts | base64 > ${TMP}/known_hosts
+    cat ~/.ssh/known_hosts | base64 -w 0 > ${TMP}/known_hosts
     set +x
     helper_text="If this step fails, make the second positional arg the path to an ssh key authenticated with Github Enterprise
     
@@ -238,9 +362,17 @@ if [[ -z ${GH_TOKEN} ]]; then
     oc create secret generic git-ssh-key --from-file=ssh-privatekey=${ssh_key} --type=kubernetes.io/ssh-auth
     helper_text=""
     oc annotate secret git-ssh-key --overwrite 'tekton.dev/git-0'='github.ibm.com'
-    oc secrets link pipeline git-ssh-key --for=mount
-    set +e
-    oc secrets unlink pipeline git-token
+    if [[ ${is_openshift} == "true" ]]; then
+        oc secrets link pipeline git-ssh-key --for=mount
+        set +e
+        oc secrets unlink pipeline git-token
+    else
+        kubectl patch serviceaccount default -p '{"secrets": [{"name": "git-ssh-key"}]}'
+        set +e
+#        kubectl patch serviceaccount default --type=json -p='[{"op": "remove", "path": "/data/mykey"}]'
+#        kubectl patch deploy/some-deployment --type=json -p='[{"op": "remove", "path": "/spec/template/spec/containers/0/ports/0"},{"op": "remove", "path": "/spec/template/spec/containers/0/ports/2"}]
+#        oc get sa default -o yaml | grep -A3 "secrets:" | awk '/git-token/ { print NR }' 
+    fi
     set -e
     extra_params="${extra_params} -p git-url=git@github.ibm.com:IBM-Data-Fabric/mesh-for-data.git -p wkc-connector-git-url=git@github.ibm.com:ngoracke/WKC-connector.git -p vault-plugin-secrets-wkc-reader-url=git@github.ibm.com:data-mesh-research/vault-plugin-secrets-wkc-reader.git"
 else
@@ -257,9 +389,14 @@ stringData:
   password: ${GH_TOKEN}
 EOH
     oc apply -f ${TMP}/git-token.yaml
-    oc secrets link pipeline git-token --for=mount
-    set +e
-    oc secrets unlink pipeline git-ssh-key
+    if [[ ${is_openshift} == "true" ]]; then
+        oc secrets link pipeline git-token --for=mount
+        set +e
+        oc secrets unlink pipeline git-ssh-key
+    else
+        kubectl patch serviceaccount default -p '{"secrets": [{"name": "git-token"}]}'
+        set +e
+    fi
     set -e
     extra_params="${extra_params} -p git-url=https://github.ibm.com/IBM-Data-Fabric/mesh-for-data.git -p wkc-connector-git-url=https://github.ibm.com/ngoracke/WKC-connector.git -p vault-plugin-secrets-wkc-reader-url=https://github.ibm.com/data-mesh-research/vault-plugin-secrets-wkc-reader.git"
 fi
@@ -309,7 +446,7 @@ set +x
 
 echo "
 # for a pre-existing PVC that will be deleted when the namespace is deleted
-tkn pipeline start build-and-deploy -w name=images-url,emptyDir="" -w name=artifacts,claimName=artifacts-pvc -w name=shared-workspace,claimName=source-pvc -p docker-hostname=image-registry.openshift-image-registry.svc:5000 -p docker-namespace=${unique_prefix} -p NAMESPACE=${unique_prefix} -p skipTests=${skip_tests} -p transfer-images-to-icr=${transfer_images_to_icr} ${extra_params} -p git-revision=pipeline"
+tkn pipeline start build-and-deploy -w name=images-url,emptyDir="" -w name=artifacts,claimName=artifacts-pvc -w name=shared-workspace,claimName=source-pvc -p docker-hostname=${image_repo} -p dockerhub-hostname=${dockerhub_hostname} -p docker-namespace=${unique_prefix} -p NAMESPACE=${unique_prefix} -p skipTests=${skip_tests} -p transfer-images-to-icr=${transfer_images_to_icr} ${extra_params} -p git-revision=pipeline"
 
 if [[ ${run_tkn} -eq 1 ]]; then
     set -x
@@ -329,6 +466,8 @@ spec:
     value: ${unique_prefix} 
   - name: docker-hostname
     value: image-registry.openshift-image-registry.svc:5000
+  - name: dockerhub-hostname
+    value: ${dockerhub_hostname}
   - name: docker-namespace
     value: ${unique_prefix} 
   - name: git-revision
